requirements.dev.txt
# development
aiohttp-devtools==1.1.2

# testing
Faker==23.1.0
coverage==7.4.1
pytest-cov==4.1.0
pylama==8.4.1
pytest==8.0.0
pytest-aiohttp==1.0.5
pytest-alembic==0.10.7
SQLAlchemy-Utils==0.41.1
pytz==2024.1

